Handling errors

+++ commands/administration/clear.js
@@ -8,6 +8,9 @@ exports.run = (bot,message,args) => {
 		message.channel.send("Você não tem permissão para executar esse comando! ``` r!setadmin (Nick) ``` para dar permissão a um usuário!");
 		return;
 	}
+
+	if (args[0] > 100 || args[0] < 2) return message.channel.send("Envie um valor entre 2 e 100!");
+
 	try{
 		message.channel.bulkDelete(args[0]).then(messages => {
 			message.channel.send(`${messages.size} mensagens deletadas!`);
